[Green] Implement RestaurantContainer
